+++ front-end/src/App.scss
@@ -1,3 +1,3 @@
 @import 'styles/reset.scss';
 @import 'styles/fonts.scss';
-@import 'styles/colors.scss';
+@import 'styles/variables.scss';
