+++ front-end/src/components/Video/index.tsx
@@ -4,22 +4,21 @@ import VideoPlayer from "./../VideoPlayer";
 
 import "./styles.scss";
 
-const Video: React.FC = () => {
-  const [video, setVideo] = useState({
-    // src: "//vjs.zencdn.net/v/oceans.mp4"
-    qualities: [
-      {
-        title: "ocean",
-        src: "//vjs.zencdn.net/v/oceans.mp4",
-        type: "video/mp4",
-      },
-    ],
-    title: "ocean",
-  } as VideoInterface);
+const videoObj = {
+  qualities: [
+    {
+      title: "ocean",
+      src: "//vjs.zencdn.net/v/oceans.mp4",
+      type: "video/mp4",
+    },
+  ],
+  title: "ocean",
+} as VideoInterface;
 
+const Video: React.FC = () => {
   return (
     <div className="video__container">
-      <VideoPlayer video={video.qualities} title={video.title} />
+      <VideoPlayer video={videoObj.qualities} title={videoObj.title} />
     </div>
   );
 };
