+++ front-end/src/components/VideoPlayer/styles.scss
@@ -1,16 +1,5 @@
 /* container */
 .video {
-	--color-white: hsl(0, 0%, 100%);
-	--color-trans: hsl(0, 0%, 100%, .25);
-	--color-black: hsl(0, 0%, 14%);
-	--color-black-transparent: hsla(0, 0%, 14%, .7);
-
-	--layout-gutter: 1rem;
-	--zindex-meta: 1;
-	--transition: all .2s ease-in-out;
-	--transition-slow: all .5s ease-in-out;
-	--scrubberSize: 16px;
-
 	font-weight: 100;
 	display: flex;
 	position: relative;
