--- front-end/src/styles/_colors.scss
@@ -1,6 +1,0 @@
-:root {
-  --background: #FFFFFF 0% 0% no-repeat padding-box;
-  --black: #2E2B30;
-  --white: #fff;
-  --gray: #8a8c90;
-}
